Update baseConfig.yml to include UHD results (#884)
* Update baseConfig.yml to include UHD results

UHD is a subset of HD, so it arguably makes sense to include 4k results when searching for HD (instead of creating a new category).

* Increase HD Movie and TV max filesize

Movie HD increased to `90G` max.
TV HD increased to `12G` max.

This may need adjusting later. Ranges taken from a quick search of $indexer.